:truck: Rename db to lib
